Add library.properties
A very popular way to install libraries is to use GitHub's **Clone or download > Download .ZIP file** link to download the library and then the Arduino IDE's **Sketch > Include Library > Add .ZIP Library** feature to install the downloaded file. The problem is that the Arduino IDE only recognizes libraries that either have a header file or a library.properties file in the root library folder. Since this library stores the source files in the src it is necessary to add a library.properties file to permit this installation method.

This change will also provide better metadata for the library's Library Manager entry.

This change would also be necessary in order to submit the library for inclusion in the Library Manager index.